+++ frontend/src/features/auth/api/register.ts
