+++ frontend/src/features/auth/components/RegisterBox.tsx
